+++ app/services/BottleBeerFactory.js
@@ -1,27 +1,20 @@
 (function() {
-  
 
-
-    var BottleBeerFactory = function($http, base64) {
+    var BottleBeerFactory = function($http, $data, base64) {
       
         var factory = {};
 
+        // Prep Work
+        var user =  $data.user;
+        var pass =  $data.pass;
+        var beg =   $data.db_beg;
+        var end =   $data.db_end;
+        var beerURL = beg + "drink/_view/BottleBeer" + end;     // Data URL (String)
+
         factory.getBottleBeer = function () {
-          
-          var user = "commeelyinguederistroust";
-          var pass = "2imaJ47lISdfKTXV48oAKdXp";
-          var db = "covebeta/_design/drink/_view/BottleBeer?include_docs=true&reduce=false";
-
-          var beerURL = "https://infiniteinnovations.cloudant.com/" + db;
-                    
-          $http.defaults.headers.common['Accept'] = 'application/json';
-                    
-          $http.defaults.headers.common['Authorization'] = 'Basic ' + base64.encode(user + ':' + pass);
-
-            /*
-            console.log("Below is the the object base64");
-            console.log(base64);
-            */
+            // Set Headers
+            $http.defaults.headers.common['Accept'] = 'application/json';
+            $http.defaults.headers.common['Authorization'] = 'Basic ' + base64.encode(user + ':' + pass);
 
             return $http.get(beerURL);
         };      
@@ -29,29 +22,8 @@
         return factory;
     };
     
-    BottleBeerFactory.$inject = ['$http', 'base64'];
+    BottleBeerFactory.$inject = ['$http', '$data', 'base64'];
         
-    angular.module('coveApp').factory('BottleBeerFactory',
-                                           BottleBeerFactory);
+    angular.module('coveApp').factory('BottleBeerFactory', BottleBeerFactory);
                                            
 }());
-/////////////////
-// OLD STUFF:
-////////////////
-
-/* FROM EXAMPLE:
-factory.getBottleBeer = function(drinkId) {
- return $http.get('/bottleBeers/' + drinkId);
-};
-*/
-/* for getting local database
-   factory.getBottleBeer = function () {
-
-           return $http.get('app/db/BottleBeer.json');
-        };
-*/
-
-// Headers and Stuff
-
-// $http.defaults.headers.common = {"Access-Control-Request-Headers": "Accept, Origin, Authorization"};
-// $http.defaults.headers.common['Origin'] = 'http://cove-admin-demo-125709.use1-2.nitrousbox.com:8888';