+++ app/values.js
@@ -1,10 +1,5 @@
 angular.module('coveApp').value('appSettings', {
     title: 'Cove App Admin',
-    version: '0.2'
+    version: '0.3'
 
 });
-
-/*angular.module('coveApp').constant('appSettings', {
-    title: 'Customers Application',
-    version: '1.0'
-});*/