+++ changelog.md
@@ -5,3 +5,9 @@
 
 ## 0.2 - 2014-06-29
 - BottleBeers view functioning correctly. Able to successfully display formatted JSON of BottleBeers.
+
+## 0.3 - 2014-07-02
+- Added Base64 object with encode/decode.
+- Using JSON data from remote Cloudant source
+- Relies on HTTP Basic Auth and CORS.
+- Cleaned up source quite a bit.